feat: restructure example folder
