+++ example/pom.xml
@@ -10,7 +10,7 @@
 	</parent>
 	<groupId>com.elevenpaths.almaraz</groupId>
 	<artifactId>almaraz-example</artifactId>
-	<version>0.2.0</version>
+	<version>0.2.1</version>
 	<name>almaraz-example</name>
 	<description>Example of a WebFlux server leveraging almaraz</description>
 
@@ -27,14 +27,13 @@
 		<dependency>
 			<groupId>com.elevenpaths.almaraz</groupId>
 			<artifactId>almaraz</artifactId>
-			<version>0.2.0</version>
+			<version>0.2.1</version>
 		</dependency>
 		<dependency>
 			<groupId>net.logstash.logback</groupId>
 			<artifactId>logstash-logback-encoder</artifactId>
 			<version>6.1</version>
 		</dependency>
-		
 		
 		<dependency>
 			<groupId>org.projectlombok</groupId>
@@ -58,6 +57,14 @@
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
+				<executions>
+					<execution>
+						<goals>
+							<goal>repackage</goal>
+							<goal>build-info</goal>
+						</goals>
+					</execution>
+				</executions>
 			</plugin>
 		</plugins>
 	</build>
