+++ example/src/main/java/com/elevenpaths/almaraz/example/Application.java
@@ -7,7 +7,7 @@ package com.elevenpaths.almaraz.example;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = { "com.elevenpaths.almaraz" })
 public class Application {
 
 	public static void main(String[] args) {
