+++ example/src/main/java/com/elevenpaths/almaraz/example/configuration/ApplicationConfiguration.java
@@ -2,13 +2,14 @@
  * Copyright (c) Telefonica I+D. All rights reserved.
  */
 
-package com.elevenpaths.almaraz.example;
+package com.elevenpaths.almaraz.example.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.info.BuildProperties;
 import org.springframework.context.annotation.Configuration;
 
 import com.elevenpaths.almaraz.AlmarazConfiguration;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
  * Spring {@link Configuration} extending {@link AlmarazConfiguration} to include the default middlewares
