+++ pom.xml
@@ -5,7 +5,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.elevenpaths.almaraz</groupId>
 	<artifactId>almaraz</artifactId>
-	<version>0.2.0</version>
+	<version>0.2.1</version>
 	<packaging>jar</packaging>
 
 	<name>${project.groupId}:${project.artifactId}</name>
@@ -48,7 +48,7 @@
 		<spring.version>5.1.7.RELEASE</spring.version>
 		<spring-boot.version>2.1.3.RELEASE</spring-boot.version>
 		<reactor.version>3.2.9.RELEASE</reactor.version>
-		<json.schema.version>1.0.12</json.schema.version>
+		<json.schema.version>1.0.19</json.schema.version>
 		<logback.version>1.2.3</logback.version>
 		<junit.version>4.12</junit.version>
 		<mockito.version>2.23.4</mockito.version>
